Moved shared TableRow type
