+++ backend/routers/datatypes/typeData.ts
@@ -7,3 +7,8 @@ export type AggregatedData<T> = Record<
   string,
   Record<string, any> & { data: T }
 >
+
+export type TableRow<TRowData> = {
+  rowId: string
+  rowData: TRowData
+}
