+++ src/api/data/tableResponses.ts
@@ -1,0 +1,4 @@
+export interface TableRow<TRowData> {
+  rowId: string
+  rowData: TRowData
+}
